+++ src/components/Post.vue
@@ -1060,6 +1060,9 @@ $reaction-padding: 5px;
 }
 
 .post-attachments {
+  display: flex;
+  flex-direction: column;
+  gap: calc($block-inner-padding / 2);
   margin: $block-inner-padding 0;
   padding: 0 $block-inner-padding;
 }