+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Make gaps between post editor blocks equal to gaps between post blocks.
+
 ## [3.20.1] - 2025-03-17
 
 ### Changed