+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [3.21.0] - 2025-03-24
+
 ### Changed
 
 - Make gaps between post editor blocks equal to gaps between post blocks.